test vehicule finit + vérification de la présence de l'objet dans le deleteVehicule (service)

+++ agricore_spring_boot/src/main/java/agricore/projet/services/VehiculeService.java
@@ -74,6 +74,11 @@ public class VehiculeService {
     }
 
     public void delete(Integer id) {
+
+        if(!daovehicule.existsById(id)) {
+            throw new VehiculeNotFound(id);
+        }
+
         daovehicule.deleteById(id);
         
     }
